+++ CHANGELOG.md
@@ -15,8 +15,10 @@ depending on the global `AssetManager` were removed.
   - `FixtureDef` builder methods supporting all shapes (`CircleShape`, `PolygonShape`, `ChainShape`, `EdgeShape`).
   - `FixtureDef.filder` extension methods simplifying `Filter` properties setup.
   - `BodyDefinition` is a `BodyDef` extension providing `Body` building DSL. Used internally by `World.body`.
+  - `FixtureDefinition` is a `FixtureDef` extension providing `Fixture` building DSL. Used internally by `BodyDefinition`.
   - `fixture`, `circle`, `box`, `polygon`, `chain`, `loop` and `edge` extension `Fixture` building methods added to `Body`.
   - `earthGravity` property allowing to set `World` gravity roughly matching Earth's gravity.
+  - `onCreate` callbacks in `BodyDefinition` and `FixtureDefinition` giving access to built `Body` and `Fixture` instances in building blocks.
 - **[CHANGE]** (`ktx-i18n`) Static `I18NBundle` instance container - `I18n` - was removed.
 - **[CHANGE]** (`ktx-i18n`) Top level `nls` functions were removed.
 - **[FEATURE]** (`ktx-i18n`) `nls` property and method added to `BundleLine` for extra readability.